integrated auth.js into all pages
