+++ public/custom_figure.html
@@ -11,6 +11,7 @@
   <script type="module" src="../src/js/pages/customFigureBuilder.js" defer></script>
   <script type="module" src="../src/js/components/customKeyFigureSidebar.js" defer></script>
   <script type="module" src="../src/script.js"></script>
+  <script type="module" src="../src/js/auth/auth.js" defer></script>
 
 </head>
 <body>
