+++ public/datenschutz.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Datenschutz</title>
   <script src="https://cdn.tailwindcss.com"></script>
+  <script type="module" src="../src/js/auth/auth.js" defer></script>
   <link rel="stylesheet" href="../src/styles/styles.css">
 </head>
 <body class="bg-gray-100 flex flex-col items-center p-4 min-h-screen">
