+++ public/index.html
@@ -11,6 +11,7 @@
   <script type="module" src="../src/js/components/companySidebar.js" defer></script>
   <script src="../src/js/keyFigureData/loadCompanyData.js" defer></script>
   <script type="module" src="../src/script.js" defer></script>
+  <script type="module" src="../src/js/auth/auth.js" defer></script>
 </head>
