+++ public/login.html
@@ -6,6 +6,7 @@
   <title>Login</title>
   <script src="https://cdn.tailwindcss.com"></script>
   <script type="module" src="../src/script.js"></script>
+  <script type="module" src="../src/js/auth/auth.js" defer></script>
   <link rel="stylesheet" href="../src/styles/styles.css">
   <link rel="stylesheet" href="../src/styles/infobox.css">
 </head>
